[E2E tests] Use config from grunt
